@@ -1,12 +1,10 @@
-﻿namespace Scotec.Web.ImageServer
+﻿namespace Scotec.Web.ImageServer;
+
+public struct ImageInfo
 {
-    public struct ImageInfo
-    {
-        public string Path { get; set; }
+    public string Path { get; set; }
 
-        public int? Width { get; set; } 
+    public int? Width { get; set; }
 
-        public int? Height { get; set; } 
-
-    }
+    public int? Height { get; set; }
 }
