@@ -9,16 +9,16 @@ public static class HttpContextExtensions
         return GetCultureFromPath(httpContext.Request.Path.Value);
     }
 
-    public static string GetCultureFromReferer(this HttpContext httpContext)
+    public static CultureInfo? GetCultureFromReferer(this HttpContext httpContext)
     {
         var referer = httpContext.Request.Headers["Referer"].ToString();
         if (string.IsNullOrEmpty(referer))
         {
-            return string.Empty;
+            return null;
         }
         var uri = new Uri(referer);
 
-        return GetCultureFromPath(uri.LocalPath);
+        return new CultureInfo(GetCultureFromPath(uri.LocalPath));
     }
 
     private static string GetCultureFromPath(string? path)
@@ -27,7 +27,8 @@ public static class HttpContextExtensions
         {
             var segments = path.Split(new[] { '/' }, StringSplitOptions.RemoveEmptyEntries);
 
-            if (segments.Length >= 1 && segments[0].Length == 2)
+            // Only 2-letter codes are currently supported.
+            if (segments is [{ Length: 2 }, ..])
             {
                 var currentCulture = segments[0];
                 return currentCulture;
