@@ -1,4 +1,4 @@
-﻿using System.Diagnostics;
+﻿using Microsoft.AspNetCore.Components.Server.Circuits;
 using Microsoft.AspNetCore.Localization;
 using Microsoft.Extensions.Options;
 
@@ -13,21 +13,25 @@ public static class RequestCultureMiddlewareExtensions
         options.RequestCultureProviders.Clear();
         options.ApplyCurrentCultureToResponseHeaders = true;
 
-        options.AddInitialRequestCultureProvider(new CustomRequestCultureProvider(async context =>
+        var requestCultureProvider = builder.ApplicationServices.GetRequiredService<UriRequestCultureProvider>();
+        options.AddInitialRequestCultureProvider(requestCultureProvider);
+
+        return builder.UseMiddleware<UrlLocalizationAwareWebSocketsMiddleware>(Options.Create(options))
+            .UseRequestLocalization(options);
+    }
+
+    public static IServiceCollection AddUrlRequestLocalization(this IServiceCollection services,
+        LanguageOptions languageOptions)
+    {
+        services.AddTransient<UriRequestCultureProvider>();
+        services.AddScoped<CircuitHandler, RequestCultureCircuitHandler>();
+        services.Configure<RequestLocalizationOptions>(options =>
         {
-            var currentCulture = context.GetCultureFromRequest();
+            options.DefaultRequestCulture = new RequestCulture(languageOptions.DefaultLanguage);
+            options.SupportedCultures = languageOptions.Languages;
+            options.SupportedUICultures = languageOptions.Languages;
+        });
 
-            var requestCulture = new ProviderCultureResult(currentCulture, currentCulture);
-
-            return await Task.FromResult(requestCulture);
-        }));
-
-        // Do not pass the options to UseMiddleware(). Instead, register them as a singleton during service registration. Passing the options as parameters
-        // to UseMiddleware() does not register the options as a service and makes them available only to the middleware.
-        // IOptions<> is registered as a singleton by default and can be injected into any service lifetime. However, if it has not been
-        // registered before, the service locator returns a default instance of the options.
-        return builder.UseMiddleware<UrlLocalizationAwareWebSocketsMiddleware>();
-        //.UseMiddleware<UrlLocalizationAwareWebSocketsMiddleware>(Options.Create(options))
-        //.UseRequestLocalization(options);
+        return services;
     }
 }
