@@ -1,9 +1,13 @@
 ﻿using Microsoft.Extensions.DependencyInjection.Extensions;
 using Microsoft.Extensions.Options;
+using Scotec.Web.ImageServer.Caching;
+using Scotec.Web.ImageServer.Processor;
+using Scotec.Web.ImageServer.Provider;
+using Scotec.Web.ImageServer.Server;
 
 namespace Scotec.Web.ImageServer;
 
-public static class ImageServerMifddlewareExtensions
+public static class ImageServerMiddlewareExtensions
 {
     public static IApplicationBuilder UseImageServer(this IApplicationBuilder builder)
     {
@@ -14,15 +18,31 @@ public static class ImageServerMifddlewareExtensions
 
     public static IServiceCollection AddImageServer(this IServiceCollection services)
     {
-        services.AddScoped<IImageServer, ImageServer>()
-                .AddScoped<IImageProcessor, MagickImageProcessor>()
-                .AddImageProvider<IImageProvider, LocalImageProvider>("images")
-                .AddSingleton<IImageCache, InMemoryImageCache>();
+        services.AddScoped<IImageServer, DefaultImageServer>()
+            .AddScoped<IImageProcessor, MagickImageProcessor>()
+            .AddSingleton<IImageCache, InMemoryImageCache>();
 
         return services;
     }
 
-    public static IServiceCollection AddImageProvider<TService, TImplementation>(this IServiceCollection services, string key) 
+    public static IServiceCollection AddLocalImageProvider(this IServiceCollection services, string folderName)
+    {
+        services.AddImageProvider<IImageProvider, LocalImageProvider>(folderName);
+
+        return services;
+    }
+
+
+    public static IServiceCollection AddAzureBlobStorageImageProvider(this IServiceCollection services,
+        string containerName)
+    {
+        services.AddImageProvider<IImageProvider, AzureBlobStorageImageProvider>(containerName);
+
+        return services;
+    }
+
+    public static IServiceCollection AddImageProvider<TService, TImplementation>(this IServiceCollection services,
+        string key)
         where TService : class, IImageProvider
         where TImplementation : class, TService
     {
