@@ -1,6 +1,9 @@
-﻿namespace Scotec.Web.Robots.Sitemap;
+﻿using System.Globalization;
+
+namespace Scotec.Web.Robots.Sitemap;
 
 public interface ISitemapEntryProvider
 {
     IEnumerable<ISitemapEntry> Entries { get; }
+    IEnumerable<ISitemapEntry> GetEntries(CultureInfo culture);
 }
